requirements.txt
django>3
djangorestframework
markdown
django-filter
djangorestframework-simplejwt